Refactor layers: Remove MomentumLayer, add DeltaLayer, MomentumLayer, and KalmanLayer
- Deleted MomentumLayer implementation from momentum.py.
- Introduced DeltaLayer in hopfield.py for associative memory with active overwriting, implementing the delta rule for state updates.
- Added MomentumLayer in jordan.py, maintaining fast temporal pattern capture with gated EMA.
- Implemented KalmanLayer in wiener.py for uncertainty-aware state estimation using diagonal Kalman filtering.
- Each new layer includes detailed documentation and follows the SYNTHE architecture for consistency.

+++ src/layers/attention.py
@@ -159,7 +159,7 @@ class AttentionProbe(SyntheLayer):
                 active_idx = activate.nonzero(as_tuple=True)[0]
                 active_x = x[active_idx]
                 active_out = self._sliding_window_attention(active_x)
-                output[active_idx] = active_out
+                output[active_idx] = active_out.to(output.dtype)
         else:
             # No confidence signal — always compute
             output = self._sliding_window_attention(x)
